+++ src/__tests__/schema.ts
@@ -2,7 +2,7 @@ import {
   DateTime,
   Default,
   Enum,
-  Index,
+  Id,
   Boolean,
   Int,
   Limit,
@@ -11,16 +11,24 @@ import {
   OneToMany,
   Unique,
   UpdatedAt,
-  Varchar,
+  String,
   ManyToOne,
   Mixin,
   Pk,
   OneToOne,
+  Map,
+  Key,
+  Float,
+  Ignore,
 } from '../';
 
 // from: https://www.prisma.io/docs/concepts/components/prisma-schema#example
 
-const Role = Enum('Role', ['USER', 'ADMIN'] as const);
+const Role = Enum(
+  'Role',
+  Key('USER', Map('user')),
+  Key('ADMIN'),
+);
 
 const Post = Model('Post');
 const User = Model('User');
@@ -29,24 +37,23 @@ const Timestamps = Mixin()
   .Field('createdAt', DateTime(Default('now()')))
   .Field('updatedAt', DateTime(UpdatedAt));
 
-// prettier-ignore
 User
-  .Field("id",          Int(Index, Default("autoincrement()")))
-  .Field("email",       Varchar(Unique))
-  .Field("name",        Varchar(Nullable))
-  .Field("role",        Role("USER"))
-  .Relation("posts",    OneToMany(Post))
-  .Field("bestPostId",  Int())
-  .Relation("bestPost", OneToOne(Post, Pk("id").Fk("bestPostId")))
+  .Field('id',          Int(Id, Default('autoincrement()'), Map('_id')))
+  .Field('email',       String(Unique))
+  .Field('name',        String(Nullable, Ignore))
+  .Field('height',      Float(Default(1.80)))
+  .Field('role',        Role('USER', Nullable))
+  .Relation('posts',    OneToMany(Post))
+  .Field('bestPostId',  Int())
+  .Relation('bestPost', OneToOne(Post, Pk('id').Fk('bestPostId')))
   .Mixin(Timestamps);
 
-// prettier-ignore
 Post
-  .Field("id",          Int(Index, Default("autoincrement()")))
-  .Field("published",   Boolean(Default(false)))
-  .Field("title",       Varchar(Limit(255)))
-  .Field("authorId",    Int(Nullable))
-  .Relation("author",   ManyToOne(User, Pk("id").Fk("authorId"), Nullable))
+  .Field('id',          Int(Id, Default('autoincrement()')))
+  .Field('published',   Boolean(Default(false)))
+  .Field('title',       String(Limit(255)))
+  .Field('authorId',    Int(Nullable))
+  .Relation('author',   ManyToOne(User, Pk('id').Fk('authorId'), Nullable))
   .Mixin(Timestamps)
   .Raw(`@@map("comments")`);