+++ src/codegen/modifiers.ts
@@ -3,8 +3,11 @@ import { Type } from '../types/types';
 import { transform } from './transform';
 
 // TODO: less shitty way of doing this
-export const modifier = (type: Type, modifier: Modifier): string => {
-  switch (modifier.type as any) {
+export const modifier = <T extends Type>(
+  type: T,
+  modifier: Modifier<T>,
+): string => {
+  switch (modifier.type) {
     case 'default':
       return `@default(${
         type == 'Enum' ? modifier.value : transform(modifier.value)
@@ -15,5 +18,9 @@ export const modifier = (type: Type, modifier: Modifier): string => {
       return '@unique';
     case 'updatedAt':
       return '@updatedAt';
+    case 'ignore':
+      return '@ignore';
+    case 'map':
+      return `@map("${modifier.value}")`;
   }
 };