+++ src/public/enum.ts
@@ -1,24 +1,22 @@
 import * as Types from '../types';
 import { nonNullable } from '../types/utils';
 
-export class Enum<K extends readonly string[]>
+export class Enum<K extends Types.Fields.EnumKey[]>
   extends Function
   implements Types.Blocks.Block<'enum'>
 {
-  type: 'enum' = 'enum';
-  columns: Types.Blocks.Model['columns'];
+  type: 'enum' = 'enum' as const;
+  columns: Types.Column<'EnumKey'>[];
 
   constructor(public name: string, keys: K) {
     super();
 
-    this.columns = keys.map(
-      k =>
-        ({
-          name: k,
-          type: 'Enum',
-          modifiers: [],
-        } as Types.Column<'Enum'>),
-    );
+    // Define the keys of the enum
+    this.columns = keys.map(k => ({
+      name: k.name,
+      type: 'EnumKey' as const,
+      modifiers: k.modifiers,
+    }));
 
     return new Proxy(this, {
       apply: (
@@ -38,18 +36,18 @@ export class Enum<K extends readonly string[]>
     ...modifiers: Types.Modifier<'Enum'>[]
   ): Types.Fields.Field<'Enum'> {
     return {
-      type: 'Enum',
+      type: 'Enum' as const,
       modifiers: [
-        { type: 'enum', value: this.name },
+        { type: 'enum' as const, value: this.name },
         initial
-          ? { type: 'default', value: initial }
+          ? { type: 'default' as const, value: initial }
           : initial === null
-          ? { type: 'nullable', value: true }
+          ? { type: 'nullable' as const, value: true }
           : null,
         ...modifiers,
       ].filter(
         (v, i, a) => nonNullable(v) && a.findIndex(m => m.type == v.type) === i,
-      ),
-    } as Types.Fields.Field<'Enum'>;
+      ) as Types.Modifier<'Enum'>[],
+    };
   }
 }