+++ src/public/modifiers.ts
@@ -1,11 +1,18 @@
 import * as Types from '../types';
 
 export const Default = <
-  T extends Types.Fields.Primitive,
+  T extends Types.Fields.Scalar,
   K extends Types.TypeData[T]['default'],
 >(
   value: K,
 ): Types.Modifier<T, 'default'> => ({ type: 'default', value });
+
+export const Map = <
+  T extends Types.Fields.Scalar | 'EnumKey',
+  K extends Types.TypeData[T]['map'],
+>(
+  value: K,
+): Types.Modifier<T, 'map'> => ({ type: 'map', value });
 
 export const Unique = {
   type: 'unique',
@@ -22,11 +29,16 @@ export const Nullable = {
   value: true,
 } as const;
 
-export const Index = {
-  type: 'index',
+export const Id = {
+  type: 'id',
   value: true,
 } as const;
 
-export const Limit = <K extends Types.TypeData['Varchar']['limit']>(
+export const Ignore = {
+  type: 'ignore',
+  value: true,
+} as const;
+
+export const Limit = <K extends Types.TypeData['String']['limit']>(
   value: K,
-): Types.Modifier<'Varchar', 'limit'> => ({ type: 'limit', value });
+): Types.Modifier<'String', 'limit'> => ({ type: 'limit', value });