+++ src/types/mixins.ts
@@ -1,5 +1,6 @@
+import { Fields } from '.';
 import { Column } from './columns';
 
 export type Mixin = {
-  columns: Column[];
+  columns: Column<Fields.Scalar>[];
 };