+++ src/types/modifiers.ts
@@ -3,5 +3,7 @@ import { Type, TypeData } from './types';
 // Column modifiers, e.g. @default(), @nullable() etc.
 export type Modifier<
   T extends Type = Type,
-  Property extends keyof TypeData[T] = keyof TypeData[T],
+  Property extends Modifiers<T> = Modifiers<T>,
 > = { type: Property; value: TypeData[T][Property] };
+
+export type Modifiers<T extends Type> = keyof TypeData[T];